# Benchmark active indices