--- a/beacon-chain/cache/benchmarks_test.go
+++ b/beacon-chain/cache/benchmarks_test.go
@@ -4,12 +4,10 @@ import (
 	"testing"
 )
 
-var aInfo300K, _ = createIndices(300000)
-var aInfo1M, _ = createIndices(1000000)
+var indices300k, _ = createIndices(300000)
 var epoch = uint64(1)
 var _, treeIndices300K = createIndices(300000)
-var byteIndices300k = convertUint64ToByteSlice(treeIndices300K);
-
+var byteIndices300k = convertUint64ToByteSlice(treeIndices300K)
 
 func createIndices(count int) (*ActiveIndicesByEpoch, []uint64) {
 	indices := make([]uint64, 0, count)
@@ -30,7 +28,7 @@ func BenchmarkCachingAddRetrieve(b *testing.B) {
 		b.N = 10
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
-			if err := c.AddActiveIndicesList(aInfo300K); err != nil {
+			if err := c.AddActiveIndicesList(indices300k); err != nil {
 				b.Fatal(err)
 			}
 		}
@@ -91,7 +89,6 @@ func BenchmarkTreeAddRetrieve(b *testing.B) {
 func BenchmarkBloomFilter(b *testing.B) {
 	bf := NewBloomFilter()
 
-
 	b.Run("BLOOMADD300K", func(b *testing.B) {
 		b.N = 10
 		b.ResetTimer()
@@ -100,13 +97,12 @@ func BenchmarkBloomFilter(b *testing.B) {
 		}
 	})
 
-
 	b.Run("BLOOMTEST300K", func(b *testing.B) {
 		b.N = 10
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
 			if !bf.TestActiveIndicesBloomFilter(byteIndices300k) {
-				break;
+				break
 			}
 		}
 		bf.ClearBloomFilter()
@@ -116,7 +112,6 @@ func BenchmarkBloomFilter(b *testing.B) {
 func BenchmarkCFilter(b *testing.B) {
 	cf := NewCFilter()
 
-
 	b.Run("CFILTERADD300K", func(b *testing.B) {
 		b.N = 10
 		b.ResetTimer()
@@ -125,16 +120,13 @@ func BenchmarkCFilter(b *testing.B) {
 		}
 	})
 
-
 	b.Run("CFILTERLOOKUP300K", func(b *testing.B) {
 		b.N = 10
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
 			if !cf.LookupActiveIndicesCFilter(byteIndices300k) {
-				break;
+				break
 			}
 		}
 	})
 }
-
-