--- a/beacon-chain/cache/bloom_filter.go
+++ b/beacon-chain/cache/bloom_filter.go
@@ -1,18 +1,18 @@
 package cache
 
 import (
-	"sync"
 	"encoding/binary"
-	
+	"sync"
+
 	"github.com/willf/bloom"
 )
 
 // ActiveIndicesBloomFilter defines BloomFilter for storing and twsting of active indices
-// Bloom Filter is a space-efficient probabilistic data structure 
+// Bloom Filter is a space-efficient probabilistic data structure
 // that is used to test whether an element is a member of a set
 type ActiveIndicesBloomFilter struct {
 	filter *bloom.BloomFilter
-	lock sync.RWMutex
+	lock   sync.RWMutex
 }
 
 // A Bloom filter has two parameters: m - maximum size, k -the number of hashing functions on elements of the set
@@ -23,11 +23,11 @@ const (
 
 // NewBloomFiltercreates a new Bloom Filter for storing and testing active indices
 func NewBloomFilter() *ActiveIndicesBloomFilter {
-	return &ActiveIndicesBloomFilter{filter : bloom.New(m, k)}
+	return &ActiveIndicesBloomFilter{filter: bloom.New(m, k)}
 }
 
 // AddActiveIndicesBloomFilter adds a byte representation of active indices to a bloom filter
-func (bf *ActiveIndicesBloomFilter) AddActiveIndicesBloomFilter(byteIndices [][]byte)  {
+func (bf *ActiveIndicesBloomFilter) AddActiveIndicesBloomFilter(byteIndices [][]byte) {
 	bf.lock.Lock()
 	defer bf.lock.Unlock()
 	for _, i := range byteIndices {
@@ -48,7 +48,7 @@ func (bf *ActiveIndicesBloomFilter) TestActiveIndicesBloomFilter(byteIndices [][
 }
 
 // ClearBloomFilter clears keys of bloom filter
-func (bf *ActiveIndicesBloomFilter) ClearBloomFilter()  {
+func (bf *ActiveIndicesBloomFilter) ClearBloomFilter() {
 	bf.filter = bf.filter.ClearAll()
 }
 
@@ -56,27 +56,9 @@ func (bf *ActiveIndicesBloomFilter) ClearBloomFilter()  {
 func convertUint64ToByteSlice(activeIndices []uint64) [][]byte {
 	byteIndices := make([][]byte, 0, len(activeIndices))
 	for _, i := range activeIndices {
-		n1 := make([]byte,8)
-		binary.BigEndian.PutUint64(n1,i)
+		n1 := make([]byte, 8)
+		binary.BigEndian.PutUint64(n1, i)
 		byteIndices = append(byteIndices, n1)
 	}
 	return byteIndices
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-