--- a/beacon-chain/cache/tree.go
+++ b/beacon-chain/cache/tree.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/petar/GoLLRB/llrb"
 )
+
 // ActiveIndicesTree defines active validator indices in LLRB tree
 type ActiveIndicesTree struct {
 	tree *llrb.LLRB