|Resolves] #2740
[Part of] #3055

This PR adds 5 files:

1. beacon-chain/cache/tree.go that contains methods to insert and retrieve items from Left-Leaning Red-Black (LLRB) tree 

2. beacon-chain/cache/tree_test.go has methods for testing LLRB tree 

3.beacon-chain/cache/bloom_filter.go  has methods to add and make membership checks for active indices in a bloom filter. Bloom Filter is a space-efficient probabilistic data structure
that is used to test whether an element is a member of a set. 

4. beacon-chain/cache/cfilter.go has for storing and lookup of active indices for cuckoo filter.
Cuckoo filter is a Bloom filter replacement for approximated set-membership queries. Cuckoo filters support adding and removing items dynamically while achieving even higher performance than Bloom filters.

5.beacon-chain/cache/benchmarks_test.go has benchmarks for current, LLRB implementation, bloom and cuckoo filters.

**Benchmark results:**

- Current implementation:

BenchmarkCachingAddRetrieve/ADD300K-3 10 504 ns/op 16 B/op 1 allocs/op
BenchmarkCachingAddRetrieve/RETR300K-3 10 224 ns/op 0 B/op 0 allocs/op

- LLRB tree

BenchmarkTreeAddRetrieve/TREEADD300K-3 10 347157294 ns/op 3839993 B/op 329999 allocs/op
BenchmarkTreeAddRetrieve/RETRTREE300K-3 10 14174378 ns/op 2400256 B/op 1 allocs/op
BenchmarkTreeAddRetrieve/TREEADDNOREPL300K-3 10 593535169 ns/op 31199995 B/op 899999 allocs/op
BenchmarkTreeAddRetrieve/RETRTREE300K#01-3 10 201188730 ns/op 50405376 B/op 1 allocs/op

- Bloom filter 

BenchmarkBloomFilter/BLOOMADD300K-3         	      10	 108189202 ns/op	29100064 B/op	  600000 allocs/op
BenchmarkBloomFilter/BLOOMTEST300K-3        	      10	       440 ns/op	      11 B/op	       0 allocs/op

- Cuckoo filter

BenchmarkCFilter/CFILTERADD300K-3         	      10	11617697675 ns/op	 4800000 B/op	  600000 allocs/op
BenchmarkCFilter/CFILTERLOOKUP300K-3      	      10	      2255 ns/op	     152 B/op	      19 allocs/op

Conclusion : current implementation is the most efficient.