@cyberbono3 2019-08-07
there are some issues with [bls](https://travis-ci.com/prysmaticlabs/prysm/builds/122326480) so there are no issues from my side, this PR is ready for review