@cyberbono3 2019-08-14
> Can you also add numbers on how much memory each cache takes up ? It would be good to compare memory efficiency across different cache types

I guess I provided memory allocation here 567 ns/op              16 B/op          1 allocs/op byte per operand isn't it?