@cyberbono3 2019-08-18
@prestonvanloon  approved to include  benchmarks of current implementation only. All others benchmarks have been discarded.